--- a/packages/dds/matrix/.gitignore
+++ b/packages/dds/matrix/.gitignore
@@ -51,5 +51,8 @@ nyc
 intel_modules/
 temp_modules/
 
+# Output folder for memory profiling tests
+.memoryTestsOutput
+
 bench/dist/
 bench/profile.txt