--- a/packages/dds/matrix/src/test/memory/matrix.spec.ts
+++ b/packages/dds/matrix/src/test/memory/matrix.spec.ts
@@ -34,95 +34,119 @@ describe("Matrix memory usage", () => {
 		// See the comment at the top of the test suite for more details.
 	});
 
-	benchmarkMemory(
-		new (class implements IMemoryTestObject {
-			title = "Create empty Matrix";
-			minSampleCount = 500;
-
-			private localMatrix: SharedMatrix = createLocalMatrix("testLocalMatrix");
-
-			async run() {
-				this.localMatrix = createLocalMatrix("testLocalMatrix");
-			}
-		})(),
-	);
-
-	const numbersOfEntriesForTests = [100, 1000, 10_000];
-
-	numbersOfEntriesForTests.forEach((x) => {
+	describe("Detached Matrix", () => {
 		benchmarkMemory(
 			new (class implements IMemoryTestObject {
-				title = `Insert and remove ${x} columns`;
+				title = "Create empty Matrix";
+				minSampleCount = 500;
+
 				private localMatrix: SharedMatrix = createLocalMatrix("testLocalMatrix");
 
 				async run() {
-					for (let i = 0; i < x; i++) {
-						this.localMatrix.insertCols(0, 100);
-						this.localMatrix.removeCols(0, 100);
-					}
-				}
-
-				beforeIteration() {
 					this.localMatrix = createLocalMatrix("testLocalMatrix");
 				}
 			})(),
 		);
 
-		benchmarkMemory(
-			new (class implements IMemoryTestObject {
-				title = `Insert and remove ${x} rows`;
-				private localMatrix: SharedMatrix = createLocalMatrix("testLocalMatrix");
+		const numbersOfEntriesForTests = [100, 1000, 5000];
 
-				async run() {
-					for (let i = 0; i < x; i++) {
-						this.localMatrix.insertRows(0, 100);
-						this.localMatrix.removeRows(0, 100);
+		numbersOfEntriesForTests.forEach((x) => {
+			benchmarkMemory(
+				new (class implements IMemoryTestObject {
+					title = `Insert and remove a column ${x} times`;
+					private localMatrix: SharedMatrix = createLocalMatrix("testLocalMatrix");
+
+					async run() {
+						for (let i = 0; i < x; i++) {
+							this.localMatrix.insertCols(0, 1);
+							this.localMatrix.removeCols(0, 1);
+						}
 					}
-				}
 
-				beforeIteration() {
-					this.localMatrix = createLocalMatrix("testLocalMatrix");
-				}
-			})(),
-		);
-
-		benchmarkMemory(
-			new (class implements IMemoryTestObject {
-				title = `Insert and remove ${x} rows and columns`;
-				private localMatrix: SharedMatrix = createLocalMatrix("testLocalMatrix");
-
-				async run() {
-					for (let i = 0; i < x; i++) {
-						this.localMatrix.insertCols(0, 100);
-						this.localMatrix.insertRows(0, 100);
-						this.localMatrix.removeCols(0, 100);
-						this.localMatrix.removeRows(0, 100);
+					beforeIteration() {
+						this.localMatrix = createLocalMatrix("testLocalMatrix");
 					}
-				}
+				})(),
+			);
 
-				beforeIteration() {
-					this.localMatrix = createLocalMatrix("testLocalMatrix");
-				}
-			})(),
-		);
+			benchmarkMemory(
+				new (class implements IMemoryTestObject {
+					title = `Insert and remove one row ${x} times`;
+					private localMatrix: SharedMatrix = createLocalMatrix("testLocalMatrix");
 
-		benchmarkMemory(
-			new (class implements IMemoryTestObject {
-				title = `Set ${x} cells`;
-				private localMatrix = createLocalMatrix("testLocalMatrix");
-
-				async run() {
-					this.localMatrix.insertCols(0, x);
-					this.localMatrix.insertRows(0, x);
-					for (let i = 0; i < x; i++) {
-						this.localMatrix.setCell(0, i, "a");
+					async run() {
+						for (let i = 0; i < x; i++) {
+							this.localMatrix.insertRows(0, 1);
+							this.localMatrix.removeRows(0, 1);
+						}
 					}
-				}
 
-				beforeIteration() {
-					this.localMatrix = createLocalMatrix("testLocalMatrix");
-				}
-			})(),
-		);
+					beforeIteration() {
+						this.localMatrix = createLocalMatrix("testLocalMatrix");
+					}
+				})(),
+			);
+
+			benchmarkMemory(
+				new (class implements IMemoryTestObject {
+					title = `Insert and remove a row and column ${x} times`;
+					only = true;
+					private localMatrix: SharedMatrix = createLocalMatrix("testLocalMatrix");
+
+					async run() {
+						for (let i = 0; i < x; i++) {
+							this.localMatrix.insertCols(0, 1);
+							this.localMatrix.insertRows(0, 1);
+							this.localMatrix.removeCols(0, 1);
+							this.localMatrix.removeRows(0, 1);
+						}
+					}
+
+					beforeIteration() {
+						this.localMatrix = createLocalMatrix("testLocalMatrix");
+					}
+				})(),
+			);
+
+			benchmarkMemory(
+				new (class implements IMemoryTestObject {
+					title = `Set a 3-character string in ${x} cells`;
+					only = true;
+					private localMatrix = createLocalMatrix("testLocalMatrix");
+
+					async run() {
+						for (let i = 0; i < x; i++) {
+							this.localMatrix.setCell(0, i, "abc");
+						}
+					}
+
+					beforeIteration() {
+						this.localMatrix = createLocalMatrix("testLocalMatrix");
+						this.localMatrix.insertRows(0, 1);
+						this.localMatrix.insertCols(0, x);
+					}
+				})(),
+			);
+
+			benchmarkMemory(
+				new (class implements IMemoryTestObject {
+					title = `Set a number in ${x} cells`;
+					only = true;
+					private localMatrix = createLocalMatrix("testLocalMatrix");
+
+					async run() {
+						for (let i = 0; i < x; i++) {
+							this.localMatrix.setCell(0, i, 1);
+						}
+					}
+
+					beforeIteration() {
+						this.localMatrix = createLocalMatrix("testLocalMatrix");
+						this.localMatrix.insertRows(0, 1);
+						this.localMatrix.insertCols(0, x);
+					}
+				})(),
+			);
+		});
 	});
 });