## Description

- Update the perf unit tests for SharedMatrix to be more accurate in what they do VS what they say they do.
- Decrease the maximum number of repetitions for some tests. It seems it legitimately takes a significant amount of time to complete those repetitions, to the point where the actual test runtime plus overhead of the benchmarking infrastructure sometimes hits the mocha timeout of 90s in CI.
- Exclude the output files of the memory perf tests from source control.

In the screenshot below we can see each iteration of the longer tests took ~700ms, and the perf benchmarking package does 50 iterations at a minimum to try to produce statistically significant measurements. That only accounts for 35s out of the 90s that we see the tests take in CI sometimes, but that could vary depending on the particular environment where the tests are running.

![image](https://user-images.githubusercontent.com/716334/222836678-dc4dd472-2d0f-4a78-877c-b569a533cb19.png)

Addresses [AB#2963](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2963) .